REF: move callback fix to dbr.make_callback
